Fixed tile rendering subpixel math

+++ TwelveEngine/Game2D/ScreenSpace.cs
@@ -4,7 +4,7 @@
         public float Y;
         public float Width;
         public float Height;
-        public float TileSize;
+        public int TileSize;
         public (float x,float y) getCenter() {
             return (X + Width / 2, Y + Height / 2);
         }
